refactor: move utils folder into test folder
